adicionando middeware em todas rotas

+++ resources/views/products/index.blade.php
@@ -1,23 +1,23 @@
 <x-layout title="Produtos">
     <a href="/products/create" class="btn btn-dark"> Cadastrar produto </a>
-    {{-- Message --}}
-    @if (Session::has('success'))
-        <div class="alert alert-success alert-dismissible" role="alert">
-            <button type="button" class="close" data-dismiss="alert">
-                <i class="fa fa-times"></i>
-            </button>
-            <strong>Successo!</strong> {{ session('success') }}
-        </div>
-    @endif
+     {{-- Message --}}
+        @if (Session::has('success'))
+            <div class="alert alert-success alert-dismissible" role="alert">
+                <button type="button" class="close" data-dismiss="alert">
+                    <i class="fa fa-times"></i>
+                </button>
+                <strong>Successo!</strong> {{ session('success') }}
+            </div>
+        @endif
 
-    @if (Session::has('error'))
-        <div class="alert alert-danger alert-dismissible" role="alert">
-            <button type="button" class="close" data-dismiss="alert">
-                <i class="fa fa-times"></i>
-            </button>
-            <strong>Erro!</strong> {{ session('error') }}
-        </div>
-    @endif
+        @if (Session::has('error'))
+            <div class="alert alert-danger alert-dismissible" role="alert">
+                <button type="button" class="close" data-dismiss="alert">
+                    <i class="fa fa-times"></i>
+                </button>
+                <strong>Erro!</strong> {{ session('error') }}
+            </div>
+        @endif
 
     <table class="table table-dark">
         <thead>
@@ -42,7 +42,7 @@
                     </a>
                 </td>
                 <td class="text-right" style="width: 80px">
-                    <a class="btn btn-warning" href="/products/delete/{{$product->id}}" style="color:rgb(0, 0, 0); padding: 2px">
+                    <a class="btn btn-warning" onclick="return confirm('Deseja excluir esse item?')" href="/products/delete/{{$product->id}}" style="color:rgb(0, 0, 0); padding: 2px">
                         excluir
                     </a>
                 </td>
@@ -50,11 +50,6 @@
             @endforeach
         </tbody>
     </table>
-
-
-
-
-
     <div class="d-flex justify-content-center">
         {{ $products->links()}}
     </div>
